requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.18.1
idna==2.10
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.25.10
